+++ nengo/linear_system.py
@@ -3,7 +3,7 @@ import numpy as np
 from nengo._vendor.scipy.signal import cont2discrete, ss2tf, tf2ss, tf2zpk, zpk2ss
 from nengo.base import Process
 from nengo.exceptions import ValidationError
-from nengo.params import BoolParam, EnumParam, NdarrayParam
+from nengo.params import BoolParam, EnumParam, NdarrayParam, Parameter
 from nengo.rc import rc
 
 
@@ -367,3 +367,11 @@ class LinearSystem(Process):
                 return Y
 
         return step_linearsystem
+
+
+class LinearSystemParam(Parameter):
+    equatable = True
+
+    def coerce(self, instance, system):  # pylint: disable=arguments-renamed
+        self.check_type(instance, system, LinearSystem)
+        return super().coerce(instance, system)